refactor: 255以上の値があるので,uint8からuint16にした

+++ jpeg2voxel_converter.py
@@ -18,7 +18,7 @@ for filename in sorted(os.listdir(slice_folder_path)):
 # ボクセルデータを作成する
 depth = len(slice_images)
 # 3はR,G,Bの3つのチャンネルということ
-voxels = np.zeros((width, height, depth, 3), dtype=np.uint8)
+voxels = np.zeros((width, height, depth, 3), dtype=np.uint16)
 
 # 各スライス画像を処理し、ボクセルデータに変換する
 for z, slice_image in enumerate(slice_images):
